requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.8
et-xmlfile==1.1.0
idna==3.3
loguru==0.5.3
numpy==1.21.4
openai==0.11.0
openpyxl==3.0.9
pandas==1.3.4
pandas-stubs==1.2.0.39
PyAudio==0.2.11
pycodestyle==2.8.0
pyobjc==8.0
pyobjc-core==8.0
pyobjc-framework-Accessibility==8.0
pyobjc-framework-Accounts==8.0
pyobjc-framework-AddressBook==8.0
pyobjc-framework-AdServices==8.0
pyobjc-framework-AdSupport==8.0
pyobjc-framework-AppleScriptKit==8.0
pyobjc-framework-AppleScriptObjC==8.0
pyobjc-framework-ApplicationServices==8.0
pyobjc-framework-AppTrackingTransparency==8.0
pyobjc-framework-AudioVideoBridging==8.0
pyobjc-framework-AuthenticationServices==8.0
pyobjc-framework-AutomaticAssessmentConfiguration==8.0
pyobjc-framework-Automator==8.0
pyobjc-framework-AVFoundation==8.0
pyobjc-framework-AVKit==8.0
pyobjc-framework-BusinessChat==8.0
pyobjc-framework-CalendarStore==8.0
pyobjc-framework-CallKit==8.0
pyobjc-framework-CFNetwork==8.0
pyobjc-framework-ClassKit==8.0
pyobjc-framework-CloudKit==8.0
pyobjc-framework-Cocoa==8.0
pyobjc-framework-Collaboration==8.0
pyobjc-framework-ColorSync==8.0
pyobjc-framework-Contacts==8.0
pyobjc-framework-ContactsUI==8.0
pyobjc-framework-CoreAudio==8.0
pyobjc-framework-CoreAudioKit==8.0
pyobjc-framework-CoreBluetooth==8.0
pyobjc-framework-CoreData==8.0
pyobjc-framework-CoreHaptics==8.0
pyobjc-framework-CoreLocation==8.0
pyobjc-framework-CoreMedia==8.0
pyobjc-framework-CoreMediaIO==8.0
pyobjc-framework-CoreMIDI==8.0
pyobjc-framework-CoreML==8.0
pyobjc-framework-CoreMotion==8.0
pyobjc-framework-CoreServices==8.0
pyobjc-framework-CoreSpotlight==8.0
pyobjc-framework-CoreText==8.0
pyobjc-framework-CoreWLAN==8.0
pyobjc-framework-CryptoTokenKit==8.0
pyobjc-framework-DeviceCheck==8.0
pyobjc-framework-DictionaryServices==8.0
pyobjc-framework-DiscRecording==8.0
pyobjc-framework-DiscRecordingUI==8.0
pyobjc-framework-DiskArbitration==8.0
pyobjc-framework-DVDPlayback==8.0
pyobjc-framework-EventKit==8.0
pyobjc-framework-ExceptionHandling==8.0
pyobjc-framework-ExecutionPolicy==8.0
pyobjc-framework-ExternalAccessory==8.0
pyobjc-framework-FileProvider==8.0
pyobjc-framework-FileProviderUI==8.0
pyobjc-framework-FinderSync==8.0
pyobjc-framework-FSEvents==8.0
pyobjc-framework-GameCenter==8.0
pyobjc-framework-GameController==8.0
pyobjc-framework-GameKit==8.0
pyobjc-framework-GameplayKit==8.0
pyobjc-framework-ImageCaptureCore==8.0
pyobjc-framework-IMServicePlugIn==8.0
pyobjc-framework-InputMethodKit==8.0
pyobjc-framework-InstallerPlugins==8.0
pyobjc-framework-InstantMessage==8.0
pyobjc-framework-Intents==8.0
pyobjc-framework-IOSurface==8.0
pyobjc-framework-iTunesLibrary==8.0
pyobjc-framework-KernelManagement==8.0
pyobjc-framework-LatentSemanticMapping==8.0
pyobjc-framework-LaunchServices==8.0
pyobjc-framework-libdispatch==8.0
pyobjc-framework-LinkPresentation==8.0
pyobjc-framework-LocalAuthentication==8.0
pyobjc-framework-MapKit==8.0
pyobjc-framework-MediaAccessibility==8.0
pyobjc-framework-MediaLibrary==8.0
pyobjc-framework-MediaPlayer==8.0
pyobjc-framework-MediaToolbox==8.0
pyobjc-framework-Metal==8.0
pyobjc-framework-MetalKit==8.0
pyobjc-framework-MetalPerformanceShaders==8.0
pyobjc-framework-MetalPerformanceShadersGraph==8.0
pyobjc-framework-MLCompute==8.0
pyobjc-framework-ModelIO==8.0
pyobjc-framework-MultipeerConnectivity==8.0
pyobjc-framework-NaturalLanguage==8.0
pyobjc-framework-NetFS==8.0
pyobjc-framework-Network==8.0
pyobjc-framework-NetworkExtension==8.0
pyobjc-framework-NotificationCenter==8.0
pyobjc-framework-OpenDirectory==8.0
pyobjc-framework-OSAKit==8.0
pyobjc-framework-OSLog==8.0
pyobjc-framework-PassKit==8.0
pyobjc-framework-PencilKit==8.0
pyobjc-framework-Photos==8.0
pyobjc-framework-PhotosUI==8.0
pyobjc-framework-PreferencePanes==8.0
pyobjc-framework-PushKit==8.0
pyobjc-framework-Quartz==8.0
pyobjc-framework-QuickLookThumbnailing==8.0
pyobjc-framework-ReplayKit==8.0
pyobjc-framework-SafariServices==8.0
pyobjc-framework-SceneKit==8.0
pyobjc-framework-ScreenSaver==8.0
pyobjc-framework-ScreenTime==8.0
pyobjc-framework-ScriptingBridge==8.0
pyobjc-framework-SearchKit==8.0
pyobjc-framework-Security==8.0
pyobjc-framework-SecurityFoundation==8.0
pyobjc-framework-SecurityInterface==8.0
pyobjc-framework-ServiceManagement==8.0
pyobjc-framework-Social==8.0
pyobjc-framework-SoundAnalysis==8.0
pyobjc-framework-Speech==8.0
pyobjc-framework-SpriteKit==8.0
pyobjc-framework-StoreKit==8.0
pyobjc-framework-SyncServices==8.0
pyobjc-framework-SystemConfiguration==8.0
pyobjc-framework-SystemExtensions==8.0
pyobjc-framework-UniformTypeIdentifiers==8.0
pyobjc-framework-UserNotifications==8.0
pyobjc-framework-UserNotificationsUI==8.0
pyobjc-framework-VideoSubscriberAccount==8.0
pyobjc-framework-VideoToolbox==8.0
pyobjc-framework-Virtualization==8.0
pyobjc-framework-Vision==8.0
pyobjc-framework-WebKit==8.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-103/six-1.15.0-py2.py3-none-any.whl
SpeechRecognition==3.8.1
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.7
